Changed up the background-color-change checks a bit - to account for all cases now. Fixed an issue where a recurring letter in userInput was not being checked against the random word properly. Added the feature to highlight all relevant buttons even after the winning word is entered.
Added CSS to include some GIFs with 'pop-in' and 'pop-out' animations when certain conditions are met, and to reset them as user starts re-entering letters.
Added CSS to make the result fade in and out and replace the heading.
Added a reset function and linked it to a game-restart-button so that user can continue with more turns of a refreshed-word-game without refreshing the page - as they do it, certain other elements on the page reset as well.
Added a score counter to track points as user progresses through game attempts.